+++ apps/cli/vitest.config.js
@@ -3,6 +3,11 @@ import { defineConfig } from "vitest/config";
 
 export default defineConfig({
   test: {
+    server: {
+      deps: {
+        inline: ["replace-in-file"],
+      },
+    },
     watch: false,
   },
 });