+++ infra/repository/environments.tf
@@ -7,3 +7,13 @@ resource "github_repository_environment" "automated_tests" {
     custom_branch_policies = false
   }
 }
+
+resource "github_repository_environment" "npm_prod_cd" {
+  environment = "npm-prod-cd"
+  repository  = module.github_repository.name
+
+  deployment_branch_policy {
+    protected_branches     = true
+    custom_branch_policies = false
+  }
+}